abctokz_repo/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "abctokz"
version = "0.1.0"
description = "A from-scratch multilingual tokenizer for English and Devanagari scripts (Hindi, Marathi, Sindhi)"
authors = [{ name = "abctokz contributors" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Text Processing :: Linguistic",
]
dependencies = [
  "typer>=0.12",
  "pydantic>=2.7",
  "regex>=2024.0",
  "rich>=13.0",
  "tqdm>=4.66",
  "numpy>=1.26",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
adapters = [
  "tokenizers>=0.19",
  "sentencepiece>=0.2",
  "transformers>=4.40",
]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "pytest-xdist>=3.5",
  "hypothesis>=6.100",
  "ruff>=0.4",
  "mypy>=1.10",
  "types-PyYAML",
  "types-tqdm",
  "hatch>=1.12",
]
all = ["abctokz[adapters,dev]"]

[project.scripts]
abctokz = "abctokz.cli.main:app"

[project.urls]
Homepage = "https://github.com/example/abctokz"
Documentation = "https://github.com/example/abctokz/docs"
Repository = "https://github.com/example/abctokz"

[tool.hatch.build.targets.wheel]
packages = ["src/abctokz"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ANN", "B", "SIM", "RUF"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN", "S"]
"examples/**" = ["ANN", "T201"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
no_implicit_reexport = true
namespace_packages = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = ["regex.*", "sentencepiece.*", "tokenizers.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--strict-markers -q"
markers = [
  "unit: unit tests",
  "integration: integration tests",
  "golden: golden tests with fixed outputs",
  "property: property-based tests",
  "slow: tests that take more than a few seconds",
]

[tool.coverage.run]
source = ["abctokz"]
branch = true

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise NotImplementedError",
  "if TYPE_CHECKING:",
  "@abstractmethod",
]
